requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
boto3==1.26.22
botocore==1.29.22
Django==4.1.3
django-filter==22.1
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
